==> 4_plot_generation.Rmd <==
---
title: "4_plot_generation"
author: "Thomas E. Metherell"
date: "23/09/2021"
output: 
  html_document: 
    keep_md: yes
---

This script generates combined plots for our manuscript on digital exclusion and adolescent mental health during the COVID-19 pandemic in the United Kingdom.

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Dependencies
We first load the necessary dependencies.

```{r dependencies}
library(tidyverse)
library(magrittr)
library(lavaan)
library(survey)
library(lavaan.survey)
library(scales)
library(cowplot)
library(egg)
library(semTable)
```

# Plot generation

```{r comp_access, echo = FALSE, results = "asis"}
# Reading model fit object without covariates
fit_comp_nocov <- readRDS(file = "Data/Objects/comp_nocov.rds")

# Tabulating model output
vlabs <- c("W9" = "Main study wave 9", "CV4" = "COVID-19 wave 4", "CV6" = "COVID-19 wave 6", "CV8" = "COVID-19 wave 8", "i" = "Intercept", "s" = "Linear coefficient", "q" = "Quadratic coefficient", "sex_merged" = "Sex", "birthy" = "Birth year", "ethn_caucasian" = "Dichotomised ethnicity", "mean_income" = "Mean monthly household income")

semTable(fit_comp_nocov, columns = c("estse", "p"), paramSets = c("latentmeans", "latentvariances", "latentcovariances", "residualvariances"), fits = c("chisq", "cfi", "rmsea"), varLabels = vlabs, file = "Data/Objects/comp_nocov_table.csv", type = "csv", print.results = FALSE)

# Preparing LGM functions for plotting
pT_cnocov <- parTable(fit_comp_nocov)
fit_comp_fn0_nocov <- function(x) pT_cnocov[pT_cnocov$lhs == "i" & pT_cnocov$op == "~1" & pT_cnocov$group == 2,]$est + pT_cnocov[pT_cnocov$lhs == "s" & pT_cnocov$op == "~1" & pT_cnocov$group == 2,]$est*x + pT_cnocov[pT_cnocov$lhs == "q" & pT_cnocov$op == "~1" & pT_cnocov$group == 2,]$est*x^2
fit_comp_fn1_nocov <- function(x) pT_cnocov[pT_cnocov$lhs == "i" & pT_cnocov$op == "~1" & pT_cnocov$group == 1,]$est + pT_cnocov[pT_cnocov$lhs == "s" & pT_cnocov$op == "~1" & pT_cnocov$group == 1,]$est*x + pT_cnocov[pT_cnocov$lhs == "q" & pT_cnocov$op == "~1" & pT_cnocov$group == 1,]$est*x^2

comp_1 <- ggplot() +
  annotate(geom = "rect", xmin = 5/6, xmax = 3.5, ymin = -Inf, ymax = Inf, fill = "grey", alpha = 0.3) +
  annotate(geom = "text", x = 3.45, y = 9.3, label = "COVID-19 pandemic", hjust = "right", vjust = "top") +
  geom_function(mapping = aes(colour = "No"), fun = fit_comp_fn0_nocov, xlim = c(0, 3), size = 1) +
  geom_function(mapping = aes(colour = "Yes"), fun = fit_comp_fn1_nocov, xlim = c(0, 3), size = 1) +
  scale_x_continuous(limits = c(0, 3.5), breaks = 0:3, labels = str_wrap(c("2017-19", "July 2020", "November 2020", "March 2021"), width = 8)) +
  ylim(9, 19) +
  xlab(NULL) +
  ylab("Total Difficulties score") +
  scale_colour_manual(values = c("No" = "#F8766D", "Yes" = "#00BFC4")) +
  theme_classic()

# Reading model fit object with covariates
fit_comp_cov <- readRDS(file = "Data/Objects/comp_cov.rds")

# Tabulating model output
semTable(fit_comp_cov, columns = c("estse", "p"), paramSets = c("latentmeans", "latentvariances", "latentcovariances", "residualvariances"), fits = c("chisq", "cfi", "rmsea"), varLabels = vlabs, file = "Data/Objects/comp_cov_table.csv", type = "csv", print.results = FALSE)

# Preparing LGM functions for plotting
pT_ccov <- parTable(fit_comp_cov)
fit_comp_fn0_cov <- function(x) pT_ccov[pT_ccov$lhs == "i" & pT_ccov$op == "~1" & pT_ccov$group == 2,]$est + pT_ccov[pT_ccov$lhs == "s" & pT_ccov$op == "~1" & pT_ccov$group == 2,]$est*x + pT_ccov[pT_ccov$lhs == "q" & pT_ccov$op == "~1" & pT_ccov$group == 2,]$est*x^2
fit_comp_fn1_cov <- function(x) pT_ccov[pT_ccov$lhs == "i" & pT_ccov$op == "~1" & pT_ccov$group == 1,]$est + pT_ccov[pT_ccov$lhs == "s" & pT_ccov$op == "~1" & pT_ccov$group == 1,]$est*x + pT_ccov[pT_ccov$lhs == "q" & pT_ccov$op == "~1" & pT_ccov$group == 1,]$est*x^2

comp_2 <- ggplot() +
  annotate(geom = "rect", xmin = 5/6, xmax = 3.5, ymin = -Inf, ymax = Inf, fill = "grey", alpha = 0.3) +
  annotate(geom = "text", x = 3.45, y = 9.3, label = "COVID-19 pandemic", hjust = "right", vjust = "top") +
  geom_function(mapping = aes(colour = "No"), fun = fit_comp_fn0_cov, xlim = c(0, 3), size = 1) +
  geom_function(mapping = aes(colour = "Yes"), fun = fit_comp_fn1_cov, xlim = c(0, 3), size = 1) +
  scale_x_continuous(limits = c(0, 3.5), breaks = 0:3, labels = str_wrap(c("2017-19", "July 2020", "November 2020", "March 2021"), width = 8)) +
  ylim(9, 19) +
  xlab(NULL) +
  ylab("Total Difficulties score") +
  labs(colour = str_wrap("Has access to a computer", width = 8)) +
  scale_colour_manual(values = c("No" = "#F8766D", "Yes" = "#00BFC4")) +
  theme_classic()

p_comp <- ggarrange(
  comp_1 + theme(legend.position = "none", panel.background = element_blank()),
  comp_2 + theme(axis.title.y = element_text(colour = NA), axis.text.y = element_blank(), panel.background = element_blank()),
  labels = c("A", "B"),
  nrow = 1
)
```

```{r int_access, echo = FALSE, results = "asis"}
# Reading model fit object without covariates
fit_int_nocov <- readRDS(file = "Data/Objects/int_nocov.rds")

semTable(fit_int_nocov, columns = c("estse", "p"), paramSets = c("latentmeans", "latentvariances", "latentcovariances", "residualvariances"), fits = c("chisq", "cfi", "rmsea"), varLabels = vlabs, file = "Data/Objects/int_nocov_table.csv", type = "csv", print.results = FALSE)

# Preparing LGM functions for plotting
pT_inocov <- parTable(fit_int_nocov)
fit_int_fn0_nocov <- function(x) pT_inocov[pT_inocov$lhs == "i" & pT_inocov$op == "~1" & pT_inocov$group == 2,]$est + pT_inocov[pT_inocov$lhs == "s" & pT_inocov$op == "~1" & pT_inocov$group == 2,]$est*x + pT_inocov[pT_inocov$lhs == "q" & pT_inocov$op == "~1" & pT_inocov$group == 2,]$est*x^2
fit_int_fn1_nocov <- function(x) pT_inocov[pT_inocov$lhs == "i" & pT_inocov$op == "~1" & pT_inocov$group == 1,]$est + pT_inocov[pT_inocov$lhs == "s" & pT_inocov$op == "~1" & pT_inocov$group == 1,]$est*x + pT_inocov[pT_inocov$lhs == "q" & pT_inocov$op == "~1" & pT_inocov$group == 1,]$est*x^2

int_1 <- ggplot() +
  annotate(geom = "rect", xmin = 5/6, xmax = 3.5, ymin = -Inf, ymax = Inf, fill = "grey", alpha = 0.3) +
  annotate(geom = "text", x = 3.45, y = 10.05, label = "COVID-19 pandemic", hjust = "right", vjust = "top") +
  geom_function(mapping = aes(colour = "No"), fun = fit_int_fn0_nocov, xlim = c(0, 3), size = 1) +
  geom_function(mapping = aes(colour = "Yes"), fun = fit_int_fn1_nocov, xlim = c(0, 3), size = 1) +
  scale_x_continuous(limits = c(0, 3.5), breaks = 0:3, labels = str_wrap(c("2017-19", "July 2020", "November 2020", "March 2021"), width = 8)) +
  ylim(10, 12) +
  xlab(NULL) +
  ylab("Total Difficulties score") +
  labs(colour = str_wrap("Has access to a good internet connection", width = 10), fill = NULL) +
  scale_colour_manual(values = c("No" = "#F8766D", "Yes" = "#00BFC4")) +
  theme_classic()

# Reading model fit object with covariates
fit_int_cov <- readRDS(file = "Data/Objects/int_cov.rds")

semTable(fit_int_cov, columns = c("estse", "p"), paramSets = c("latentmeans", "latentvariances", "latentcovariances", "residualvariances"), fits = c("chisq", "cfi", "rmsea"), varLabels = vlabs, file = "Data/Objects/int_cov_table.csv", type = "csv", print.results = FALSE)

# Preparing LGM functions for plotting
pT_icov <- parTable(fit_int_cov)
fit_int_fn0_cov <- function(x) pT_icov[pT_icov$lhs == "i" & pT_icov$op == "~1" & pT_icov$group == 2,]$est + pT_icov[pT_icov$lhs == "s" & pT_icov$op == "~1" & pT_icov$group == 2,]$est*x + pT_icov[pT_icov$lhs == "q" & pT_icov$op == "~1" & pT_icov$group == 2,]$est*x^2
fit_int_fn1_cov <- function(x) pT_icov[pT_icov$lhs == "i" & pT_icov$op == "~1" & pT_icov$group == 1,]$est + pT_icov[pT_icov$lhs == "s" & pT_icov$op == "~1" & pT_icov$group == 1,]$est*x + pT_icov[pT_icov$lhs == "q" & pT_icov$op == "~1" & pT_icov$group == 1,]$est*x^2

int_2 <- ggplot() +
  annotate(geom = "rect", xmin = 5/6, xmax = 3.5, ymin = -Inf, ymax = Inf, fill = "grey", alpha = 0.3) +
  annotate(geom = "text", x = 3.45, y = 10.05, label = "COVID-19 pandemic", hjust = "right", vjust = "top") +
  geom_function(mapping = aes(colour = "No"), fun = fit_int_fn0_cov, xlim = c(0, 3), size = 1) +
  geom_function(mapping = aes(colour = "Yes"), fun = fit_int_fn1_cov, xlim = c(0, 3), size = 1) +
  scale_x_continuous(limits = c(0, 3.5), breaks = 0:3, labels = str_wrap(c("2017-19", "July 2020", "November 2020", "March 2021"), width = 8)) +
  ylim(10, 12) +
  xlab(NULL) +
  ylab("Total Difficulties score") +
  labs(colour = str_wrap("Has access to a good internet connection", width = 10)) +
  scale_colour_manual(values = c("No" = "#F8766D", "Yes" = "#00BFC4")) +
  theme_classic()

p_int <- ggarrange(
  int_1 + theme(legend.position = "none", panel.background = element_blank()),
  int_2 + theme(axis.title.y = element_text(colour = NA), axis.text.y = element_blank(), panel.background = element_blank()),
  labels = c("C", "D"),
  nrow = 1
)

p3 <- plot_grid(
  p_comp, p_int,
  ncol = 1
)

print(p3)

ggsave("Figures/combined_lgm.png", height = 17, width = 15.9, units = "cm")
ggsave("Figures/combined_lgm.svg", height = 17, width = 15.9, units = "cm")
```

```{r indiv_no, echo = FALSE, results = "asis"}
# Using LGM parameters to predict individual trajectories
indiv_traj_p_cov <- predict(fit_comp_cov)

# Extracting predicted individual trajectory data
indiv_traj0_ccov <- data.frame(indiv_traj_p_cov[["0"]])
i <- nrow(indiv_traj0_ccov)

indivN_1 <- ggplot() +
  annotate(geom = "rect", xmin = 5/6, xmax = 3.5, ymin = -Inf, ymax = Inf, fill = "grey", alpha = 0.3) +
  annotate(geom = "text", x = 3.45, y = -2, label = "COVID-19 pandemic", hjust = "right", vjust = "top") +
  lapply(1:i, function(i) geom_function(fun = function(x) indiv_traj0_ccov$i[i] + indiv_traj0_ccov$s[i] * x + indiv_traj0_ccov$q[i] * x^2, mapping = aes(colour = "No"), xlim = c(0, 3), size = 0.5, alpha = 20/nrow(indiv_traj0_ccov))) +
  geom_function(mapping = aes(colour = "No"), fun = fit_comp_fn0_cov, xlim = c(0, 3), size = 2) +
  scale_x_continuous(limits = c(0, 3.5), breaks = 0:3, labels = str_wrap(c("2017-19", "July 2020", "November 2020", "March 2021"), width = 8)) +
  ylim(-2, 32) +
  xlab(NULL) +
  ylab("Total Difficulties score") +
  scale_colour_manual(values = c("No" = "#F8766D")) +
  theme_classic()

# Using LGM parameters to predict individual trajectories
indiv_traj_p <- predict(fit_int_cov)

# Extracting predicted individual trajectory data
indiv_traj0_icov <- data.frame(indiv_traj_p[["0"]])
i <- nrow(indiv_traj0_icov)

indivN_2 <- ggplot() +
  annotate(geom = "rect", xmin = 5/6, xmax = 3.5, ymin = -Inf, ymax = Inf, fill = "grey", alpha = 0.3) +
  annotate(geom = "text", x = 3.45, y = -2, label = "COVID-19 pandemic", hjust = "right", vjust = "top") +
  lapply(1:i, function(i) geom_function(fun = function(x) indiv_traj0_icov$i[i] + indiv_traj0_icov$s[i] * x + indiv_traj0_icov$q[i] * x^2, mapping = aes(colour = "No"), xlim = c(0, 3), size = 0.5, alpha = 20/nrow(indiv_traj0_icov))) +
  geom_function(mapping = aes(colour = "No"), fun = fit_int_fn0_cov, xlim = c(0, 3), size = 2) +
  scale_x_continuous(limits = c(0, 3.5), breaks = 0:3, labels = str_wrap(c("2017-19", "July 2020", "November 2020", "March 2021"), width = 8)) +
  ylim(-2, 32) +
  xlab(NULL) +
  ylab("Total Difficulties score") +
  scale_colour_manual(values = c("No" = "#F8766D")) +
  theme_classic()

p4 <- ggarrange(
  indivN_1 + theme(legend.position = "none", panel.background = element_blank()),
  indivN_2 + theme(legend.position = "none", axis.title.y = element_text(colour = NA), axis.text.y = element_blank(), panel.background = element_blank()),
  labels = c("A", "B"),
  nrow = 1
)

ggsave("Figures/lims.png", plot = p4)
ggsave("Figures/lims.svg", plot = p4)
```
